Add varlen handling for borrowing some variable length data space from key
Except for adding a parameter to rb_create, this should be no change for the
case that varlen is not used, that is, taglen = zero. The intention is, caller
will source include with taglen = 0 or 1. Caller will take care of adjusting
data and key lengths according to difference between fixed reclen and actual
record len, which is allowed to be up to 254 bytes more than reclen.