Move new record block initialilzation to caller because bigmap does not know how to do it
This became an issue when polymorphic record block format appeared
on the map. Before that, a callback form bigmap worked fine, but now
this is awkward, especailly because bigmap is in C and cannot pass
through the necessary class information.

Maybe bigmap should not be in C, but that is another story. For now
it works.