Make shardmap.h standalone by inlining bigmap.h and a portion of pmem.h
This is a bit awkward because bigmap.h has to be hand-merged and parts
of pmem.h now appear in two separate source files. But shardmap.h now is
standalone as required for a library. The situation could or should be
improved a bit by hiding all details of pmem from shardmap.h.